A bit of refactoring
Anything starts with a small hack and eventually becomes a big hack. This is on that way. I've changed the filtering because just the table was filtered by location, not the rest. Now everything is filtered. Closes #5 which I created on the fly.

A few legacy functions erased also from utils-node.js